+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Support rendering on pixmaps larger than 8191x8191 pixels.
   From now, `Pixmap` is limited only by the amount of memory caller has.
+- `Transform::map_points`
 
 ## [0.6.0] - 2021-08-21
 ### Added